fix column datatype,add debuging prints and swap data arrays

+++ sor.h
@@ -37,7 +37,7 @@ extern "C" {
     }sor;
 
     sor* init_sor(int rank, int num_proc, int matrix_width,
-                int matrix_height, int p_h, int p_w, float p_threshold);
+                int matrix_height, float p_h, float p_w, float p_threshold);
     int compute(sor* block, int cnv_check);
     float compute_red(sor* block, int cnv_check);
     float compute_black(sor* block, int cnv_check);
